requirements.txt
Faker==19.3.1
Flask==3.0.0
python-ulid==1.1.0
